pyproject.toml
[project]
name = "finalcoursework"
version = "0.1.0"
description = ""
authors = [
    {name = "Haohanmaiyami",email = "kharitonovayan2018@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.8.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "coverage (>=7.10.6,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"

